UserTalkTool: *rewrote JS to remove unwanted Prototype dependency, moved it into its own file and loaded it via ResourceLoader *introduced new 'usertalktool' user right and gave it to sysops by default instead of checking for the removed User::isSysop() method in the special page's execute() method *coding style tweaks *removed some unused variables *changed wfDebug calls to wfDebugLog calls *more and better i18n *added missing image file
TODO/FIXME: added a FIXME comment in the JS file, because right now the JS both depends on the user's interface language (must be English at the moment) and the format of MediaWiki:Usertalktool-done. There should be a better way to check if sending the message was successful or not.