AuthorEmailNotification: *rewrote JS to be object-oriented and loaded it via ResourceLoader *added images/ directory and moved two images from extensions/wikihow there, they were only used by this extension *added SQL schema + update.php handler code *added new AuthorEmailNotificationHooks class and moved all hooked functions there *added missing i18n messages and renamed all messages from aen_some_message to aen-some-message, as per MediaWiki's coding conventions *added maintenance/ directory and three rewritten maintenance scripts into it, though from what I gather only emailNotifications.php is/should be used, the other two are older scripts *moved some methods only used by maintenance scripts into the appropriate script class, out of the AuthorEmailNotification class *changed wfDebug calls to wfDebugLog calls *corrected static/non-static method calls; either a method is static and is called as such or isn't, and is referred via $this or $someClass, not both... *coding style tweaks *added and improved docs
TODO FIXME #1: in AuthorEmailNotification::notifyMod(), the code checks the edit summary against the string "Reverted edits by" to identify if the edit was a rollback, but 1) this isn't the case for international (=non-English) wikis nor 2) the case when MediaWiki:Revertpage has been customized on an English wiki.
TODO FIXME #2: AuthorEmailNotification::sendModEmail() has the same flaw, it also checks for the edit summary and whether it contains one of the predefined strings. Not scalable nor internationally compatible in any possible way.
TODO CHECKME #1: I have no idea how the HTML e-mails will behave, especially now that they're essentially a strange mix of HTML and wikitext. They're probably broken a bit and thus need fixing.