AuthorLeaderboard: *rename all i18n messages to use hyphens and to include the extension name as the prefix and add a few missing/new messages to the i18n file *renamed the special page to CamelCase *added ResourceLoader support *hopefully made the #REDIRECT detection code a bit more robust
TODO FIXME: despite that I removed the awfully icky $wgLanguageCode if loop in execute(), this special page is still very English-specific as it looks for a certain English string in the edit summary of pages. That needs to be generalized so that this extension can be used on non-English wikis, too.